refactor: make dependency value more intuitive
Need to remove language from some keys now. Adding it to the dependency
value means that to_s makes more sense than the made up to_value.
The trade off is slightly more characters saved in redis, but that
could eventually be replaced by an id lookup if it actually became an
issue.