refactor: dependency storage in redis
This is a followup to previous commits that standarized the to_s
representation of dependencies as keys. Now all dependencies for any
language go under 'dependencies', so they can be deserialized straight
from the keys in the list.

The tests are also fixed after breaking in e0c5db4, where to_h was
replaced with to_hash. Instead of updating the method called,
App::Dependencies now delegates empty? to its internal hash (which makes
those tests cleaner as well).