refactor: convert rowClick to Link, improve table
The onRowClick made it difficult to highlight text in a row (namely the
repository path or the library name). The Link implementation helps
alleviate this, and its a similar pattern to how Github does links in
the code browser.

The types on the table were also improved. The goal was to make two
discriminate types where the field is a key of the Row (Basic), or
renderCell must be defined (Custom). The types were added as a Union,
but I was not able to get Typescript to narrow the type to the Basic
type when renderCell was undefined. Hopefully a future Typescript (or
my improved knowledge) can make the narrowing happen without having to
be explicit.

I also made renderCell an actual FunctionComponent instead of a
faux-component. This requires renderCell to return JSX instead of there
being a possibility to return strings, but it makes more sense type wise.