fix: latest dependabot compatibility
There were a couple issues here that I resolved, as well as some future
problems that were created.

Native helpers are installed properly now, so both bundler v1 and v2
should be parsable. This required running the build scripts for the
helpers (which require bash), and setting the dependabot path variable.

Because of the way the build script works, .bundle is now the path for
app dependencies as well. To stop the folder from leaking to the host,
.bundle was added as a special volume in compose.

The future problem that needs solving is running tests. The dependabot
build scripts configure bundler to ignore the test group, but this also
affects the app bundle config. This creates an issue where tests cannot
be run unless the configuration is removed, which requires bundle
install to be rerun in the helper folders. I think ideally the native
helper config doesn't leak outside of them being run, but I will have to
think about this more.