fix: vulnerability updates changing app updated_at
When I added the vulnerability check, every parse would store the
dependencies' vulnerability statuses. However, this caused the app to
get touched on every parse, because the touch is dependent on any redis
command run during the multi. This commit moves the vulnerability status
updates out of the redis multi so that they don't affect whether the app
gets touched.