fix: permission issues running vite dev
The notused volume ended up being a little weird because vite was
writing cache files to it. To fix this, the volume is removed, the
Dockerfile is updated to ensure that the node user owns everything, and
vite configuration is updated to write cache files in the parent
node_modules.