fix: models didn't actually type check
Tried flipping skipLibCheck to true to see if anything had changed
around swr and preact type compatibility and discovered that my
models.d.ts file wasn't actually valid. Changed .d.ts -> .ts so that it
doesn't require skipLibCheck: true to check.

Added the missing Host type, and removed old types from "v1" of the api.
Also removed all of the content in dashboard since it was using one of
the old types but the component isn't being used

Also changed the projects component to not transform the data
uneccesarily, and instead just filtering on the original field.